requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_s3
aws_cdk.aws_s3_notifications
aws_cdk.aws_lambda
aws_cdk.aws_dynamodb
aws_cdk.aws_iam